feat(Shared With You): remove unnecessary SharedWithYouCarouselCell
